% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignClusters.R
\name{alignClusters}
\alias{alignClusters}
\title{Align cluster label vectors or cluster label matrices as needed}
\usage{
alignClusters(Z1, Z2, type = NULL)
}
\arguments{
\item{Z1}{either a vector or a matrix}

\item{Z2}{either a vector or a matrix}

\item{type}{(optional) string: are Z1 & Z2 "vec"tors or "mat"rices? (auto)}
}
\value{
either a vector or a matrix, depending upon the value of "type"
}
\description{
Align cluster label vectors or cluster label matrices as needed
}

