% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignMatrices.R
\name{alignMatrices}
\alias{alignMatrices}
\title{Align matrices of cluster labels.}
\usage{
alignMatrices(Z1, Z2)
}
\arguments{
\item{Z1}{A matrix.}

\item{Z2}{A matrix of the same dimensions as Z1}
}
\value{
A matrix of the same dimensions as Z2.
}
\description{
Align matrices of cluster labels.
}

