% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignVectors.R
\name{alignVectors}
\alias{alignVectors}
\title{Align vectors of cluster labels.}
\usage{
alignVectors(Z1, Z2, ...)
}
\arguments{
\item{Z1}{A vector}

\item{Z2}{A vector of the same length as Z1}
}
\value{
A vector of the same length as Z2
}
\description{
Align vectors of cluster labels.
}
