% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaStar.R
\name{alphaStar}
\alias{alphaStar}
\title{compute mean adjusted adherence, the recommended figure of merit for bayesCC}
\usage{
alphaStar(Results)
}
\arguments{
\item{Results}{a list (the output from one completed run of bayesCC(...))}
}
\value{
a vector of mean-adjusted adherences for the run, a.k.a. alphaStar
}
\description{
compute mean adjusted adherence, the recommended figure of merit for bayesCC
}
\examples{

# see ?bayesCC
data(Results)
alphaStarDist <- data.frame(lapply(Results, alphaStar))
boxplot(alphaStarDist, main="Mean-adjusted adherence by number of clusters")

# add a violin plot
library(vioplot)
cols <- c("tomato","springgreen","darkorange","yellow")
for(i in seq_along(alphaStarDist)) {
  vioplot(alphaStarDist[[i]], col=cols[i], at=i, add=TRUE)   
}

}
