% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCAData.R
\name{BRCAData}
\alias{BRCAData}
\title{Data from TCGA BRCA paper, as used in Lock & Dunson, Bioinformatics 2013.}
\format{
a list of matrices
}
\source{
\url{https://tcga-data.nci.nih.gov/docs/publications/brca_2012/}
}
\description{
The matrices are as follows:
}
\details{
\itemize{
  \item X1. mRNA expression (HiSeq)
  \item X2. DNA methylation (HM27)
  \item X3. miRNA expression (HiSeq)
  \item X4. protein expression (RPPA) 
}
}
\seealso{
the raw files in inst/extdata and the script in demo/
}
