% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointNMF.R
\name{jointNMF}
\alias{jointNMF}
\title{attempt a joint nonnegative matrix factorization of two matrices,
as described in Wang et al. 2014, doi:10.1093/bioinformatics/btu679}
\usage{
jointNMF(mat1, mat2, findK = FALSE, howNA = c("both", "column", "row"),
  viaCV = FALSE, fracNA = 0.2)
}
\arguments{
\item{mat1}{the first matrix}

\item{mat2}{the second matrix}

\item{findK}{try to find the rank K for each matrix?}

\item{howNA}{for rank finding, add NAs column-wise, row-wise, or both?}

\item{viaCV}{for rank finding, should five-fold CV be used when imputing?}

\item{fracNA}{for rank finding, what fraction of the data should be NA'ed?}
}
\description{
attempt a joint nonnegative matrix factorization of two matrices,
as described in Wang et al. 2014, doi:10.1093/bioinformatics/btu679
}

