% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSum.R
\name{logSum}
\alias{logSum}
\title{convenience function}
\usage{
logSum(l)
}
\arguments{
\item{l}{number, vector, or matrix to be log-sum-exp'ed}
}
\value{
a numeric value
}
\description{
convenience function
}

